Extract YAML import-report annotation to importer_yaml_annotation
Sprint 4, PR #3.  First extraction from modules/importer.py
(2027 lines).  Peels off the ~256-line YAML-report annotation
cluster, which is self-contained and cohesive.

Public API in modules/importer_yaml_annotation:
- annotate_yaml_with_report(raw_text, report_lines, binary=False)
    Takes the original config.yml text plus a list of
    'imported: path' / 'not imported: path :: reason' report
    lines produced by prepare_import_payload, walks the YAML
    line-by-line, and returns an annotated copy with per-line
    ' # imported', ' # not imported (reason)' comments that
    Quickstart shows the user in the import preview.

Nine internal '_' helpers (_parse_report_details,
_parse_report_statuses, _lookup_report_reason,
_format_report_status, _build_prefix_flags, _status_from_flags,
_split_inline_comment, _parse_mapping_key,
_append_status_annotation) are ONLY called by
annotate_yaml_with_report -- verified by grep.

This is a PURE MOVE.  No annotation logic changed.  Every
annotated-line comment is byte-identical.

Compatibility:
- All 10 names are re-exported from modules/importer so
  external callers using 'importer.annotate_yaml_with_report'
  and tests using 'importer._parse_report_details' (which
  tests/test_template_gap_analyzer.py monkeypatches) keep
  working with zero changes.

- The two functions that STAY in importer.py -- sanitize_config_name
  and load_yaml_config -- keep their public locations since
  they're semantically 'config loading' not 'report annotation'.

- ImportReport class stays in importer.py because the mega
  prepare_import_payload function populates it and needs it
  in the same module.

Dependency direction: importer -> importer_yaml_annotation
(one way).  The annotation module has zero dependencies on
importer, so no cycles possible.

Impact:
- modules/importer.py: 2027 -> 1789 (-238 / -11.7%)
- modules/importer_yaml_annotation.py: NEW 285 lines

All 1284 tests pass.  Ruff + black clean.  Smoke-tested
annotate_yaml_with_report end-to-end -- produces identical
' # mapped' / ' # not imported' annotations.
